05_logging_monitoring/func03_linuxContainer/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
opencensus-extension-azure-functions
opencensus-ext-requests